Avoid circular import of `Beam3d` class
