+++ optbeam/_3d/beams.py
@@ -15,32 +15,7 @@ if not cython.compiled:
     print("\nPlease consider compiling `%s.py` via Cython: "
           "`$ cythonize -3 -i %s.py`\n" % (__name__, __name__))
 
-from types import MappingProxyType
-
-from ._helpers import _complex_dblquad
-
-
-class Beam3d:
-    """Abstract base class."""
-
-    def __init__(self, x, params, called=False):
-        """..."""
-        self.x = x   # TODO: rename x to x_shift
-        self._k = params['k']
-        self._params = MappingProxyType(params)  # read-only view of a dict
-        self.called = called
-
-    @property
-    def params(self):
-        """Beam specific parameters.
-
-        This is a read-only property.
-        """
-        return self._params
-
-    def _integrand(self, x, y):
-        """Integrand function over two coordinates x and y."""
-        raise NotImplementedError
+from ._helpers import _complex_dblquad, Beam3d
 
 
 class Beam3dSpherical(Beam3d):
